improved refresh token update logic | other minor improvements

+++ internal/repository/entities/product.go
@@ -1,10 +1,10 @@
 package entities
 
 type Product struct {
-    ID uint32 `gorm:"primaryKey;autoIncrement"`
+    ID uint64 `gorm:"primaryKey;autoIncrement"`
     Name string `gorm:"type:varchar(100);not null"`
-    Price uint32 `gorm:"not null"`
+    Price uint64 `gorm:"not null"`
     Description string `gorm:"type:varchar(800);not null"`
-    Owner uint32 
+    Owner uint64 
     User User `gorm:"foreignKey:Owner;references:ID;constraint:OnUpdate:CASCADE,OnDelete:SET NULL;"`
 }
