+++ internal/repository/entities/product/repository.go
@@ -8,8 +8,8 @@ import (
 
 type Repository interface {
 	Save(product *entities.Product) error
-	FindByID(ID *uint32, product *entities.Product) error
-	DeleteByID(ID *uint32) error
+	FindByID(ID *uint64, product *entities.Product) error
+	DeleteByID(ID *uint64) error
 }
 
 type repository struct {
@@ -24,10 +24,10 @@ func (r *repository) Save(product *entities.Product) error {
 	return r.DB.Save(product).Error
 }
 
-func (r *repository) FindByID(ID *uint32, product *entities.Product) error {
+func (r *repository) FindByID(ID *uint64, product *entities.Product) error {
 	return r.DB.Find(product, "id = ?", ID).Error
 }
 
-func (r *repository) DeleteByID(ID *uint32) error {
+func (r *repository) DeleteByID(ID *uint64) error {
 	return r.DB.Unscoped().Delete(&entities.Product{}, "id = ?", ID).Error
 }
