+++ internal/repository/entities/refreshToken.go
@@ -3,7 +3,7 @@ package entities
 import "time"
 
 type RefreshToken struct {
-	ID uint32 `gorm:"type:bigint;primaryKey;not null"`
+	ID uint64 `gorm:"type:bigint;primaryKey;not null"`
 	Token string `gorm:"type:varchar(1000);not null"`
 	Expired time.Time `gorm:"index:idx_refresh_token_expiration;type:timestamp;not null"`
 }
