+++ internal/repository/entities/unverifiedUser.go
@@ -3,7 +3,7 @@ package entities
 import "time"
 
 type UnverifiedUser struct {
-	ID uint32 `gorm:"type:bigint;primaryKey;autoIncrement"`
+	ID uint64 `gorm:"type:bigint;primaryKey;autoIncrement"`
 	Username string `gorm:"type:varchar(60);not null"`
 	Email string `gorm:"type:varchar(80);not null"`
 	Password []byte `gorm:"type:varchar(1000);not null"`
