+++ internal/repository/entities/unverifiedUser/repository.go
@@ -9,12 +9,12 @@ import (
 
 type Repository interface {
 	DeleteExpiredUsers(now time.Time)
-	Create(user *entities.UnverifiedUser) (*uint32, error)
+	Create(user *entities.UnverifiedUser) (*uint64, error)
 	IsUsernameOrEmailAlreadyInUse(username *string, email *string) bool
-	FindByID(ID *uint32, user *entities.UnverifiedUser) error
-	DeleteByID(ID *uint32) error
-	UpdateVerificationCode(ID *uint32, code *string) error
-	FindCodeByID(ID *uint32, code *string) error
+	FindByID(ID *uint64, user *entities.UnverifiedUser) error
+	DeleteByID(ID *uint64) error
+	UpdateVerificationCode(ID *uint64, code *string) error
+	FindCodeByID(ID *uint64, code *string) error
 }
 
 type repository struct {
@@ -29,7 +29,7 @@ func (r *repository) DeleteExpiredUsers(now time.Time) {
 	r.DB.Delete(&entities.UnverifiedUser{}, "expired < ?", now)
 }
 
-func (r *repository) Create(user *entities.UnverifiedUser) (*uint32, error) {
+func (r *repository) Create(user *entities.UnverifiedUser) (*uint64, error) {
 	if err := r.DB.Create(user).Error; err != nil {
 		return nil, err
 	}
@@ -45,18 +45,18 @@ func (r *repository) IsUsernameOrEmailAlreadyInUse(username *string, email *stri
 	return count > 0
 }
 
-func (r *repository) FindByID(ID *uint32, user *entities.UnverifiedUser) error {
+func (r *repository) FindByID(ID *uint64, user *entities.UnverifiedUser) error {
 	return r.DB.Find(user, "id = ?", ID).Error
 }
 
-func (r *repository) DeleteByID(ID *uint32) error {
+func (r *repository) DeleteByID(ID *uint64) error {
 	return r.DB.Delete(&entities.UnverifiedUser{}, " id = ?", ID).Error
 }
 
-func (r *repository) UpdateVerificationCode(ID *uint32, newCode *string) error {
+func (r *repository) UpdateVerificationCode(ID *uint64, newCode *string) error {
 	return r.DB.Model(&entities.UnverifiedUser{}).Where("id = ?", ID).Update("code", &newCode).Error
 }
 
-func (r *repository) FindCodeByID(ID *uint32, code *string) error {
+func (r *repository) FindCodeByID(ID *uint64, code *string) error {
 	return r.DB.Model(&entities.UnverifiedUser{}).Where("id = ?", ID).Select("code").Scan(code).Error
 }
