+++ internal/service/jwt/jwt.go
@@ -10,7 +10,7 @@ import (
 )
 
 type Manager interface {
-	Create(userID *uint32, username *string, duration time.Duration) (*string, error)
+	Create(userID *uint64, username *string, duration time.Duration) (*string, error)
 	Validate(tokenStr *string) (*UserClaims, error)
 	ValidateWithoutExpiration(tokenStr *string) (*UserClaims, error)
 }
@@ -31,7 +31,7 @@ func Init() Manager {
 	}
 }
 
-func (m *manager) Create(userID *uint32, email *string, duration time.Duration) (*string, error) {
+func (m *manager) Create(userID *uint64, email *string, duration time.Duration) (*string, error) {
 	unsignedToken := jwt.NewWithClaims(
 		jwt.SigningMethodES256,
 		InitUserClaims(userID, email, &duration),
